Refactor save
