+++ kDrive/UI/Controller/Files/MultipleSelectionFloatingPanelViewController.swift
@@ -172,7 +172,7 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
             } else {
                 for file in files {
                     if file.isDownloaded {
-                        save(file: file)
+                        FileActionsHelper.save(file: file, from: self)
                     } else {
                         downloadInProgress = true
                         collectionView.reloadItems(at: [indexPath])
@@ -180,7 +180,7 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
                         DownloadQueue.instance.observeFileDownloaded(self, fileId: file.id) { [unowned self] _, error in
                             if error == nil {
                                 DispatchQueue.main.async {
-                                    save(file: file)
+                                    FileActionsHelper.save(file: file, from: self)
                                 }
                             } else {
                                 success = false
@@ -259,47 +259,6 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
             MatomoUtils.trackBulkEvent(eventWithCategory: matomoCategory, name: "color_folder", numberOfItems: numberOfFiles)
         default:
             break
-        }
-    }
-
-    internal func save(file: File) {
-        switch file.convertedType {
-        case .image:
-            if let image = UIImage(contentsOfFile: file.localUrl.path) {
-                Task {
-                    do {
-                        try await PhotoLibrarySaver.instance.save(image: image)
-                        DispatchQueue.main.async {
-                            UIConstants.showSnackBar(message: KDriveResourcesStrings.Localizable.snackbarImageSavedConfirmation)
-                        }
-                    } catch {
-                        DDLogError("Cannot save image: \(error)")
-                        DispatchQueue.main.async {
-                            UIConstants.showSnackBar(message: KDriveResourcesStrings.Localizable.errorSave)
-                        }
-                    }
-                }
-            }
-        case .video:
-            Task {
-                do {
-                    try await PhotoLibrarySaver.instance.save(videoUrl: file.localUrl)
-                    DispatchQueue.main.async {
-                        UIConstants.showSnackBar(message: KDriveResourcesStrings.Localizable.snackbarVideoSavedConfirmation)
-                    }
-                } catch {
-                    DDLogError("Cannot save video: \(error)")
-                    DispatchQueue.main.async {
-                        UIConstants.showSnackBar(message: KDriveResourcesStrings.Localizable.errorSave)
-                    }
-                }
-            }
-        case .folder:
-            let documentExportViewController = UIDocumentPickerViewController(url: file.temporaryUrl, in: .exportToService)
-            present(documentExportViewController, animated: true)
-        default:
-            let documentExportViewController = UIDocumentPickerViewController(url: file.localUrl, in: .exportToService)
-            present(documentExportViewController, animated: true)
         }
     }
